base-layout: expose sysctl configuration to the host
There are certain `sysctl` tunables that can influence the
peformance of the solution, and as such, it is worth to make them
accessible for fine-tuning.

As part of this change, factor out the IPV6-related settings to
make it conditional on the actual configuration settings.  Also,
adjust some of the default settings to reflect saner values.

Suggested by:	Anton Saietskii <vsasjason@gmail.com>