Enable SMP for the Linux kernels.
Even though the number of CPUs made available to the guest could
be increased, they are not utilized because the different kernel
versions do not have Symmetric Multiprocessing (SMP) enabled.
This could be a limiting factor on higher load and create a
bottleneck in attaining higher throughput rates.

Change the configuration of both the long-term and the stable
kernel to support SMP systems.

Suggested by:	Anton Saietskii <vsasjason@gmail.com>