Revert 1b300a88.
Even if there is a single client, it may happen that the previous
lease has not yet expired and on the next occasion, the same IP
address is not available.  It looks awkward for now, but it is
better to let addresses allocated from a pool instead to avoid
this problem.