rtw89: Roll back to snapshot of 20240106 due to regression.
The 20240310 snapshot of `rtw89` does not work for 8852BE cards.
Similar problems were reported upstream for older Ubuntu versions
and as a fix it was suggested to move back to an older version.

Although the kernel that Wifibox/Alpine uses is much more recent
than one could find in Ubuntu 20.04, the same resolution still
works.  Based on that, chances are likely that the scope of the
regression in the `rtw89` driver is wider than assumed.

The proposed version was verified to be working, so go back to
and stick to it until the issue is resolved in upstream.

References:
- https://github.com/lwfinger/rtw89/issues/320
- https://github.com/lwfinger/rtw89/issues/323