man: add hints about the availability of hardware support
Some of the Linux drivers depend on the presence of specific
firmware files, of which might not be available in the guest image.
Make an explicit note about this and mark devices in the list of
supported hardware to highlight their respective dependencies.

Answers https://github.com/pgj/freebsd-wifibox/discussions/102