man: Restructure contents to make guest-host locations clear
Make it explicit where the configuration files are located on the
host and their relationship to the files on the guest file system.
This is to clarify certain vague references in the documentation
and help users to understand better how the Alpine-based image
works.

Fixes https://github.com/pgj/freebsd-wifibox/issues/103