aports: send logs from `wpa_supplicant` / `hostapd` to `syslogd`
Log events from WPA Supplicant and hostapd are lost because they
are run as a daemon and the output is not captured anywhere.
Use the `-s` flag to send the events directly to `syslogd` who
could then record the messages in `/var/log/messages`.  This can
facilitate troubleshooting without the need to enter the console.

Fixes freebsd-wifibox#156 [1]

[1] https://github.com/pgj/freebsd-wifibox/issues/156