rtw89: Unbreak build and update to 20250808 snapshot.
The previous GitHub repository from where the `rtw89` drivers were
consumed [1] has had its history rewritten and the old commit
reference does not work any more.  After studying the related
changes, it turned out that the support there has become limited
to 6.6 only.

For a more recent version of the driver, another repository [2]
was suggested, so switch over to that and unbreak the build.  As
a consequence, the driver is now updated to a more recent snapshot.

[1] https://github.com/a5a5aa555oo/rtw89
[2] https://github.com/morrownr/rtw89