Use GitHub Actions to build packages.
As an improvement over testing packages for building, teach GitHub
how to build and publish them incrementally.  Currently, that is
not automated but manually triggered.  The publication happens to
my default GitHub Pages site, from where the resulting packages
could be just grabbed by the `net/wifibox-alpnine` port.

There is support for building and publishing the packages from and
to branches different from the standard `main`, so that it becomes
possible to produce experimental versions.  That is what `unstable`
stands for in the default configuration parameters.